refactor(RxTest): AND-1195 Consolidate all RxTest classes to one
