+++ app/src/test/java/piuk/blockchain/android/data/datamanagers/QrCodeDataManagerTest.java
@@ -13,7 +13,7 @@ import java.util.UUID;
 import io.reactivex.observers.TestObserver;
 import piuk.blockchain.android.BlockchainTestApplication;
 import piuk.blockchain.android.BuildConfig;
-import piuk.blockchain.android.RxTest;
+import piuk.blockchain.android.testutils.RxTest;
 
 import static junit.framework.Assert.assertNotNull;
 import static junit.framework.TestCase.assertEquals;
