+++ app/src/test/java/piuk/blockchain/android/data/payments/PaymentServiceTest.kt
@@ -27,7 +27,7 @@ import org.mockito.Mockito.mock
 import org.mockito.Mockito.never
 import org.mockito.Mockito.verify
 import org.mockito.Mockito.verifyNoMoreInteractions
-import piuk.blockchain.android.RxTest
+import piuk.blockchain.android.testutils.RxTest
 import piuk.blockchain.androidcore.data.api.EnvironmentConfig
 import retrofit2.Call
 import retrofit2.Response
