+++ app/src/test/java/piuk/blockchain/android/data/payments/SendDataManagerTest.kt
@@ -14,7 +14,7 @@ import org.bitcoinj.params.BitcoinMainNetParams
 import org.junit.Before
 import org.junit.Ignore
 import org.junit.Test
-import piuk.blockchain.android.RxTest
+import piuk.blockchain.android.testutils.RxTest
 import piuk.blockchain.androidcore.data.rxjava.RxBus
 import java.math.BigInteger
