+++ app/src/test/java/piuk/blockchain/android/data/stores/DefaultFetchStrategyTest.kt
@@ -7,7 +7,7 @@ import io.reactivex.Observable
 import org.amshove.kluent.`should equal to`
 import org.amshove.kluent.mock
 import org.junit.Test
-import piuk.blockchain.android.RxTest
+import piuk.blockchain.android.testutils.RxTest
 import piuk.blockchain.androidcore.data.datastores.persistentstore.DefaultFetchStrategy
 import piuk.blockchain.androidcore.data.datastores.persistentstore.PersistentStore
 import piuk.blockchain.androidcore.utils.Optional
