+++ app/src/test/java/piuk/blockchain/android/ui/auth/PasswordRequiredPresenterTest.java
@@ -15,7 +15,7 @@ import io.reactivex.Completable;
 import io.reactivex.Observable;
 import okhttp3.MediaType;
 import okhttp3.ResponseBody;
-import piuk.blockchain.android.RxTest;
+import piuk.blockchain.android.testutils.RxTest;
 import piuk.blockchain.android.ui.launcher.LauncherActivity;
 import piuk.blockchain.android.util.DialogButtonCallback;
 import piuk.blockchain.androidbuysell.datamanagers.BuyDataManager;
