+++ app/src/test/java/piuk/blockchain/android/ui/buysell/coinify/signup/verifyemail/CoinifyVerifyEmailPresenterTest.kt
@@ -12,7 +12,7 @@ import io.reactivex.Completable
 import io.reactivex.Observable
 import org.junit.Before
 import org.junit.Test
-import piuk.blockchain.android.RxTest
+import piuk.blockchain.android.testutils.RxTest
 import piuk.blockchain.android.util.StringUtils
 import piuk.blockchain.androidbuysell.datamanagers.CoinifyDataManager
 import piuk.blockchain.androidbuysell.models.ExchangeData
