+++ app/src/test/java/piuk/blockchain/android/ui/shapeshift/inprogress/TradeInProgressPresenterTest.kt
@@ -14,7 +14,7 @@ import org.amshove.kluent.mock
 import org.junit.Before
 import org.junit.Test
 import piuk.blockchain.android.R
-import piuk.blockchain.android.RxTest
+import piuk.blockchain.android.testutils.RxTest
 import piuk.blockchain.android.ui.shapeshift.models.TradeProgressUiState
 import piuk.blockchain.androidcore.data.shapeshift.ShapeShiftDataManager
 import java.util.concurrent.TimeUnit
