+++ buysell/build.gradle
@@ -71,4 +71,5 @@ dependencies {
     testImplementation(Libraries.mockitoKotlin) {
         ['kotlin-stdlib', 'kotlin-reflect'].each { exclude module: it }
     }
+    testImplementation project(':testutils')
 }
