+++ buysell/src/test/java/piuk/blockchain/androidbuysell/datamanagers/BuyDataManagerTest.kt
@@ -12,7 +12,7 @@ import org.junit.Before
 import org.mockito.Mockito.RETURNS_DEEP_STUBS
 import org.mockito.Mockito.verify
 import org.mockito.Mockito.verifyNoMoreInteractions
-import piuk.blockchain.androidbuysell.RxTest
+import piuk.blockchain.android.testutils.RxTest
 import piuk.blockchain.androidbuysell.models.ExchangeData
 import piuk.blockchain.androidbuysell.services.BuyConditions
 import piuk.blockchain.androidbuysell.services.ExchangeService
