+++ core/src/test/java/piuk/blockchain/androidcore/data/auth/AuthDataManagerTest.kt
@@ -17,7 +17,7 @@ import org.amshove.kluent.mock
 import org.junit.Before
 import org.junit.Test
 import org.mockito.ArgumentMatchers.anyString
-import piuk.blockchain.androidcore.RxTest
+import piuk.blockchain.android.testutils.RxTest
 import piuk.blockchain.androidcore.data.access.AccessState
 import piuk.blockchain.androidcore.utils.AESUtilWrapper
 import piuk.blockchain.androidcore.utils.PrefsUtil
