+++ core/src/test/java/piuk/blockchain/androidcore/data/charts/ChartsDataManagerTest.kt
@@ -12,7 +12,7 @@ import org.amshove.kluent.any
 import org.amshove.kluent.mock
 import org.junit.Before
 import org.junit.Test
-import piuk.blockchain.androidcore.RxTest
+import piuk.blockchain.android.testutils.RxTest
 import piuk.blockchain.androidcore.data.currency.CryptoCurrencies
 import piuk.blockchain.androidcore.data.rxjava.RxBus
