+++ core/src/test/java/piuk/blockchain/androidcore/data/contacts/ContactsDataManagerTest.kt
@@ -16,7 +16,7 @@ import org.bitcoinj.crypto.DeterministicKey
 import org.junit.Before
 import org.junit.Test
 import piuk.blockchain.android.data.contacts.models.ContactTransactionModel
-import piuk.blockchain.androidcore.RxTest
+import piuk.blockchain.android.testutils.RxTest
 import piuk.blockchain.androidcore.data.contacts.datastore.ContactsMapStore
 import piuk.blockchain.androidcore.data.contacts.datastore.PendingTransactionListStore
 import piuk.blockchain.androidcore.data.rxjava.RxBus
