+++ core/src/test/java/piuk/blockchain/androidcore/data/contacts/ContactsServiceTest.kt
@@ -12,7 +12,7 @@ import info.blockchain.wallet.metadata.data.Message
 import org.amshove.kluent.shouldEqual
 import org.junit.Before
 import org.junit.Test
-import piuk.blockchain.androidcore.RxTest
+import piuk.blockchain.android.testutils.RxTest
 import java.util.HashMap
 
 class ContactsServiceTest : RxTest() {
