+++ core/src/test/java/piuk/blockchain/androidcore/data/settings/SettingsServiceTest.kt
@@ -12,7 +12,7 @@ import org.junit.Test
 import org.mockito.Mockito.mock
 import org.mockito.Mockito.verify
 import org.mockito.Mockito.verifyNoMoreInteractions
-import piuk.blockchain.androidcore.RxTest
+import piuk.blockchain.android.testutils.RxTest
 
 class SettingsServiceTest : RxTest() {
