+++ core/src/test/java/piuk/blockchain/androidcore/data/settings/datastore/SettingsMemoryStoreTest.kt
@@ -5,7 +5,7 @@ import org.amshove.kluent.`should equal`
 import org.amshove.kluent.mock
 import org.junit.Before
 import org.junit.Test
-import piuk.blockchain.androidcore.RxTest
+import piuk.blockchain.android.testutils.RxTest
 import piuk.blockchain.androidcore.utils.Optional
 
 class SettingsMemoryStoreTest : RxTest() {
