+++ core/src/test/java/piuk/blockchain/androidcore/data/shapeshift/ShapeShiftDataManagerTest.kt
@@ -25,7 +25,7 @@ import org.amshove.kluent.`should not contain`
 import org.json.JSONException
 import org.junit.Before
 import org.junit.Test
-import piuk.blockchain.androidcore.RxTest
+import piuk.blockchain.android.testutils.RxTest
 import piuk.blockchain.androidcore.data.metadata.MetadataManager
 import piuk.blockchain.androidcore.data.rxjava.RxBus
 import piuk.blockchain.androidcore.data.shapeshift.datastore.ShapeShiftDataStore
