+++ testutils/build.gradle
@@ -17,6 +17,8 @@ android {
 }
 
 dependencies {
+    api Libraries.annotations
+
     // RxJava
     api Libraries.rxJava
     api Libraries.rxKotlin
